refactor: Tile logic + update test
